// src/app/lib/generated/models/GetClaimablesResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * AI Babe Public API
 * Lorem ipsum
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface GetClaimablesResponse
 */
export interface GetClaimablesResponse {
    /**
     * 
     * @type {boolean}
     * @memberof GetClaimablesResponse
     */
    joinedDiscord: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof GetClaimablesResponse
     */
    joinedReddit: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof GetClaimablesResponse
     */
    joinedInstagram: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof GetClaimablesResponse
     */
    claimedDailyLogin: boolean;
}

/**
 * Check if a given object implements the GetClaimablesResponse interface.
 */
export function instanceOfGetClaimablesResponse(value: object): value is GetClaimablesResponse {
    if (!('joinedDiscord' in value) || value['joinedDiscord'] === undefined) return false;
    if (!('joinedReddit' in value) || value['joinedReddit'] === undefined) return false;
    if (!('joinedInstagram' in value) || value['joinedInstagram'] === undefined) return false;
    if (!('claimedDailyLogin' in value) || value['claimedDailyLogin'] === undefined) return false;
    return true;
}

export function GetClaimablesResponseFromJSON(json: any): GetClaimablesResponse {
    return GetClaimablesResponseFromJSONTyped(json, false);
}

export function GetClaimablesResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): GetClaimablesResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'joinedDiscord': json['joined_discord'],
        'joinedReddit': json['joined_reddit'],
        'joinedInstagram': json['joined_instagram'],
        'claimedDailyLogin': json['claimed_daily_login'],
    };
}

export function GetClaimablesResponseToJSON(json: any): GetClaimablesResponse {
    return GetClaimablesResponseToJSONTyped(json, false);
}

export function GetClaimablesResponseToJSONTyped(value?: GetClaimablesResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'joined_discord': value['joinedDiscord'],
        'joined_reddit': value['joinedReddit'],
        'joined_instagram': value['joinedInstagram'],
        'claimed_daily_login': value['claimedDailyLogin'],
    };
}