// src/app/lib/generated/models/UserClaimType.ts
/* tslint:disable */
/* eslint-disable */
/**
 * AI Babe Public API
 * Lorem ipsum
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const UserClaimType = {
    Discord: 'discord',
    Reddit: 'reddit',
    Instagram: 'instagram',
    DailyLogin: 'daily_login'
} as const;
export type UserClaimType = typeof UserClaimType[keyof typeof UserClaimType];


export function instanceOfUserClaimType(value: any): boolean {
    for (const key in UserClaimType) {
        if (Object.prototype.hasOwnProperty.call(UserClaimType, key)) {
            if (UserClaimType[key as keyof typeof UserClaimType] === value) {
                return true;
            }
        }
    }
    return false;
}

export function UserClaimTypeFromJSON(json: any): UserClaimType {
    return UserClaimTypeFromJSONTyped(json, false);
}

export function UserClaimTypeFromJSONTyped(json: any, ignoreDiscriminator: boolean): UserClaimType {
    return json as UserClaimType;
}

export function UserClaimTypeToJSON(value?: UserClaimType | null): any {
    return value as any;
}

export function UserClaimTypeToJSONTyped(value: any, ignoreDiscriminator: boolean): UserClaimType {
    return value as UserClaimType;
}